🚚 move: s3 파일 common으로 이동
