+++ src/main/java/com/sparta/goodbite/common/s3/config/S3Config.java
@@ -1,4 +1,4 @@
-package com.sparta.goodbite.s3;
+package com.sparta.goodbite.common.s3.config;
 
 import com.amazonaws.auth.AWSStaticCredentialsProvider;
 import com.amazonaws.auth.BasicAWSCredentials;
